+++ docs/release-notes.rst
@@ -12,6 +12,7 @@ Version 2.0.1, 09 Mar 2020
 * Drop A1 Mediator controller and tests from back-end
 * Upgrade to Spring-Boot 2.2.4.RELEASE
 * Set the first instance as the default one
+* Add methods to create, update and delete xApp stat metric URLs
 
 Version 2.0.0, 5 Feb 2020
 --------------------------